+++ libraries/SRF05/SRF05.h
@@ -2,7 +2,7 @@
 //
 //    FILE: SRF05.h
 //  AUTHOR: Rob Tillaart
-// VERSION: 0.2.0
+// VERSION: 0.3.0
 //    DATE: 2021-05-17
 // PURPOSE: Arduino library for SRF05 distance sensor
 //     URL: https://github.com/RobTillaart/SRF05
@@ -10,7 +10,7 @@
 
 #include "Arduino.h"
 
-#define SRF05_LIB_VERSION                 (F("0.2.0"))
+#define SRF05_LIB_VERSION                 (F("0.3.0"))
 
 
 const uint8_t SRF05_MODE_SINGLE      = 0;
@@ -22,7 +22,7 @@ const uint8_t SRF05_MODE_RUN_AVERAGE = 3;
 class SRF05
 {
 public:
-  explicit SRF05(const uint8_t trigger, const uint8_t echo);
+  explicit SRF05(const uint8_t trigger, const uint8_t echo = 0);
 
 
   //  configuration
@@ -40,6 +40,9 @@ public:
   void     setModeMedian(uint8_t count);
   void     setModeRunningAverage(float alpha);
   uint8_t  getOperationalMode();
+  //  interval of average and median mode
+  void     setSampleInterval(uint16_t microSeconds = 1000);
+  uint16_t getSampleInterval();
 
 
   //  get distance
@@ -49,6 +52,7 @@ public:
   float    getMeter();        //  m
   float    getInch();         //  inch = 2.54 cm
   float    getFeet();         //  feet = 12 inch
+  float    getYards();        //  yard = 3 feet = 36 inch
 
 
   //  Experimental - calibration
@@ -65,6 +69,7 @@ public:
   //  - to be investigated.
   void     setTriggerLength(uint8_t length = 10);
   uint8_t  getTriggerLength();
+
 
   //  TIMING
   uint32_t lastTime();
@@ -83,10 +88,12 @@ private:
   float    _alpha = 1.0;
   float    _value = 0;
   float    _correctionFactor = 1;
-  uint8_t  _triggerLength    = 10;
+  uint8_t  _triggerLength    = 10;       //  microSeconds
   float    _speedOfSound     = 340;      //  15°C  0%RH  Sea level
   uint32_t _lastTime = 0;
- 
+  uint16_t _sampleInterval = 1000;       //  microSeconds
+
+
   uint32_t _read();
   void     _insertSort(uint32_t * array, uint8_t size);
 };