+++ libraries/SRF05/CHANGELOG.md
@@ -6,6 +6,16 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.3.0] - 2024-02-08
+- add support for single pin operation (echo == trigger) Kudos to RobertPHeller!
+- add **float getYards()** wrapper.
+- add **void setSampleInterval(uint16_t microSeconds = 1000)** (for average & median)
+- add **uint16_t getSampleInterval()**
+- reduced pulseIn timeout to 200000 (30+ meter)
+- minor edits.
+
+----
+
 ## [0.2.0] - 2024-01-28
 - add **float calculateSpeedOfSound(float temperature, float humidity)**
 - redo lookup table SOS temperature humidity
@@ -44,7 +54,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - minor edits
 
 ## [0.1.1] - 2021-11-18
-- update build-CI, 
+- update build-CI
 - update readme.md
 - minor edits